Limpieza, arreglillos, cosas del hud

+++ Vamoos2/Assets/Scripts/Enemigos/Boss.cs
@@ -19,7 +19,7 @@ public class Boss : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        anim.SetInteger("Hits", hits);
+        anim.SetFloat("Hits", hits);
 
         if (hits == 3)
         {
